requirements.txt
asgiref==3.5.0
djangorestframework
Django
python-dotenv
sqlparse==0.4.2
requests
tzdata==2021.5


